refactor(pieces): co-locate site-trust-claims under hero/ (Block-new gate fix)
Déplace `frontend/app/config/site-trust-claims.ts` →
`frontend/app/components/pieces/hero/site-trust-claims.ts`.

Justification structurelle :
- `SITE_TRUST_CLAIMS` est consommé UNIQUEMENT par `PiecesHeroMetaStrip` (le seul
  composant du hero trust strip). Pas de réutilisation transverse.
- Le placer sous `frontend/app/config/` (config globale site) suggérait à tort
  une portée transverse. Co-location avec le bounded context hero/ aligne
  cohésion + bounded context.
- Bénéfice CI : `frontend/app/components/**` est déjà couvert par le glob
  ownership.yaml L606 (D8 + @ak125/frontend-team). Le fichier sous `hero/` passe
  naturellement le gate Block-new (owner + domain required) sans nécessiter
  d'édition de `.spec/00-canon/repository-registry/ownership.yaml`
  (canon governance owner-only, voir
  [[feedback_ownership_yaml_owner_only_edit]]).

Migration :
- `git mv frontend/app/config/site-trust-claims.ts → hero/site-trust-claims.ts`
- `PiecesHeroMetaStrip.tsx` : import `~/config/...` → `./site-trust-claims` (relatif sibling)
- `tests/unit/site-trust-claims.test.ts` : import `~/config/...` → `~/components/pieces/hero/...`
- `tests/unit/pieces-hero-meta-strip.test.tsx` : idem
- `tests/unit/pieces-hero-trust-strip.test.tsx` : idem
- Commentaire du test SoT mis à jour pour refléter le nouveau path.

Validation : 47/47 tests passent (5 fichiers verts). `frontend/app/config/`
revient à son état pré-PR (`monitoring.ts` + `site.ts` + `visual-intent.ts`
inchangés). Aucun autre fichier ne touche `~/config/site-trust-claims` (grep
clean repo-wide).
